backend/agents.py
"""
Multi-Agent System for Customer Care
Defines all specialized agents for the customer support team
"""

# Standard library
import os

# Third-party imports
from dotenv import load_dotenv
from crewai import Agent, LLM

# Local imports
from tools import search_faq, lookup_order, log_action

# Load environment variables from .env file
load_dotenv()

# ============================================================
# LLM Configuration
# ============================================================
# Using Groq API for ultra-fast inference (1-2 second responses)
# Free tier: 30 RPM, 14,400 RPD
# Model: llama-3.1-8b-instant (fast and accurate)
llm = LLM(
    model="groq/llama-3.1-8b-instant",
    api_key=os.getenv("GROQ_API_KEY")
)

# ============================================================
# Agent Definitions
# ============================================================

# Agent 1: Greeter / Intent Classifier
greeter_agent = Agent(
    role="Greeter and Intent Classifier",
    goal="Welcome customers warmly and understand what they need help with",
    backstory="""You are the friendly first point of contact for customer support.
    Your job is to make customers feel heard and quickly identify what type of help they need:
    - Order inquiries (tracking, status, issues)
    - General questions (shipping, returns, payments)
    - Complaints or issues requiring action
    - Complex problems needing escalation
    
    You are empathetic, professional, and efficient.""",
    verbose=True,
    allow_delegation=False,
    llm=llm
)

# Agent 2: Researcher / Knowledge Retriever
researcher_agent = Agent(
    role="Knowledge Researcher",
    goal="Find accurate information from FAQs and company knowledge base",
    backstory="""You are an expert at finding information quickly and accurately.
    You have access to the company's FAQ database and can search for answers about:
    - Shipping policies and timelines
    - Return and refund procedures
    - Payment methods and billing
    - Order tracking processes
    
    You always cite your sources and admit when you don't have information.""",
    verbose=True,
    allow_delegation=False,
    tools=[search_faq],
    llm=llm
)

# Agent 3: Order Specialist
order_specialist_agent = Agent(
    role="Order Specialist",
    goal="Handle all order-related inquiries including tracking and status updates",
    backstory="""You are the go-to expert for anything related to customer orders.
    You can look up order status, tracking information, and estimated delivery dates.
    You explain order statuses clearly and set appropriate expectations.
    
    When an order isn't found, you politely ask the customer to verify the number
    or suggest alternative ways to locate their order.""",
    verbose=True,
    allow_delegation=False,
    tools=[lookup_order],
    llm=llm
)

# Agent 4: Resolver / Action Taker
resolver_agent = Agent(
    role="Problem Resolver",
    goal="Take appropriate actions to resolve customer issues",
    backstory="""You are empowered to take action to solve customer problems.
    You can:
    - Log refund requests
    - Schedule callbacks
    - Escalate to human agents when needed
    - Create support tickets
    
    You always explain what action you're taking and why.
    For sensitive actions (refunds, cancellations), you ask for confirmation first.""",
    verbose=True,
    allow_delegation=False,
    tools=[log_action],
    llm=llm
)

# Agent 5: Quality Reviewer (Reflection/Critique)
quality_reviewer_agent = Agent(
    role="Quality Assurance Reviewer",
    goal="Ensure responses are accurate, helpful, and professional before delivery",
    backstory="""You are the final checkpoint before responses go to customers.
    You review the team's proposed response and check:
    - Is it accurate and complete?
    - Does it answer all the customer's questions?
    - Is the tone appropriate and empathetic?
    - Are there any errors or missing information?
    - Should we gather more information first?
    
    You can approve responses or send them back for improvement.
    You maintain high quality standards while being efficient.""",
    verbose=True,
    allow_delegation=False,
    llm=llm
)

# Agent 6: Supervisor / Orchestrator
supervisor_agent = Agent(
    role="Team Supervisor",
    goal="Coordinate the team to efficiently resolve customer inquiries",
    backstory="""You are the experienced team lead who orchestrates the customer support team.
    You decide:
    - Which agents should handle which parts of the inquiry
    - When to involve multiple agents
    - When the issue is resolved
    - When to escalate to human support
    - When to pause for customer approval
    
    You ensure smooth handoffs between agents and maintain conversation flow.
    You prioritize customer satisfaction while being efficient.""",
    verbose=True,
    allow_delegation=True,
    llm=llm
)

# Export all agents
ALL_AGENTS = {
    "greeter": greeter_agent,
    "researcher": researcher_agent,
    "order_specialist": order_specialist_agent,
    "resolver": resolver_agent,
    "quality_reviewer": quality_reviewer_agent,
    "supervisor": supervisor_agent
}
